requirements.txt
```txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
openai==1.3.7
anthropic==0.8.1
transformers==4.36.2
torch==2.1.1
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.13
chromadb==0.4.18
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3
PyPDF2==3.0.1
python-docx==0.8.11
pyttsx3==2.90
SpeechRecognition==3.10.0
librosa==0.10.1
soundfile==0.12.1
gtts==2.4.0
youtube-dl==2021.12.17
beautifulsoup4==4.12.2
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2
werkzeug==2.3.7
gunicorn==21.2.0